Remodel item by unit and item by weight
Expanding the models to support calculating discounts on products of
basket items.

We assume that a product will have knowledge on the pricing offer.
One should the base offer, which is the basal price for a certain
number of units. Another offer would be a discount offer which has a
different price per certain number of units.

Another assumption is that the basket item should be responsible to know
what is the number of units it contains. Product should be responsible
to know its qualities, such as price offers, but basket item knows how
many of those are.

The same model is replicated for items per weight, as they have a
different unit, such as kilos.