Extends model to support calculating discount put into products
The model has been refactored in such a way that a discount can be
represented as another set of price per units. For example, 'buy one
get one free' could be interpreted as 'two for the price of one'.

We want to be able to calculate the discount in items both by weight and
unit.