Test further pricing with discount scenarios
It is important to test both items by unit and weight, that if the
quantity to be bought is more than the product discount, the remaining
units will be calculated by the product base price.

Also, another case in both kinds of items, where the quantity to be
bought is at least two times more than the product discount. In such
cases, we have to verify that a discount should be applied multiple
times.