Refactor total calculator
Basket is responsible to hold the state and provide a total for all the
items. The calculation of this total is a responsibility of the
calculator. The calculator does not need to share/hold any state from
the Basket, it could just be a collection of functions offering the
capability to get the total.